pkg/seq/lexer: merges files
